Refactor some lib/evaluate headers for sanity
