+++ lib/evaluate/call.h
@@ -19,9 +19,14 @@
 #include "type.h"
 #include "../common/indirection.h"
 #include "../parser/char-block.h"
+#include "../semantics/attr.h"
 #include <optional>
 #include <ostream>
 #include <vector>
+
+namespace Fortran::semantics {
+class Symbol;
+}
 
 namespace Fortran::evaluate {
 
@@ -37,7 +42,9 @@ struct ActualArgument {
 
   std::optional<parser::CharBlock> keyword;
   bool isAssumedRank{false};  // TODO: make into a function of the value
-  bool isAlternateReturn{false};
+  bool isAlternateReturn{false};  // when true, "value" is a label number
+
+  // TODO: Mark legacy %VAL and %REF arguments
 
   // Subtlety: There is a distinction that must be maintained here between an
   // actual argument expression that is a variable and one that is not,
@@ -48,5 +55,38 @@ struct ActualArgument {
 };
 
 using ActualArguments = std::vector<std::optional<ActualArgument>>;
+
+// Intrinsics are identified by their names and the characteristics
+// of their arguments, at least for now.
+using IntrinsicProcedure = const char *;  // not an owning pointer
+
+struct SpecificIntrinsic {
+  explicit SpecificIntrinsic(IntrinsicProcedure n) : name{n} {}
+  SpecificIntrinsic(IntrinsicProcedure n, std::optional<DynamicType> &&dt,
+      int r, semantics::Attrs a)
+    : name{n}, type{std::move(dt)}, rank{r}, attrs{a} {}
+  std::ostream &Dump(std::ostream &) const;
+
+  IntrinsicProcedure name;
+  bool isRestrictedSpecific{false};  // if true, can only call it
+  std::optional<DynamicType> type;  // absent if and only if subroutine call
+  int rank{0};
+  semantics::Attrs attrs;  // ELEMENTAL, POINTER
+};
+
+struct ProcedureDesignator {
+  EVALUATE_UNION_CLASS_BOILERPLATE(ProcedureDesignator)
+  explicit ProcedureDesignator(SpecificIntrinsic &&i) : u{std::move(i)} {}
+  explicit ProcedureDesignator(const semantics::Symbol &n) : u{&n} {}
+  std::optional<DynamicType> GetType() const;
+  int Rank() const;
+  bool IsElemental() const;
+  Expr<SubscriptInteger> LEN() const;
+  const semantics::Symbol *GetSymbol() const;
+  std::ostream &Dump(std::ostream &) const;
+
+  // TODO: When calling X%F, pass X as PASS argument unless NOPASS
+  std::variant<SpecificIntrinsic, const semantics::Symbol *> u;
+};
 }
 #endif  // FORTRAN_EVALUATE_CALL_H_
