+++ lib/evaluate/intrinsics.h
@@ -16,39 +16,17 @@
 #define FORTRAN_EVALUATE_INTRINSICS_H_
 
 #include "call.h"
-#include "type.h"
-#include "../common/idioms.h"
+#include "../parser/char-block.h"
 #include "../parser/message.h"
-#include "../semantics/attr.h"
 #include "../semantics/default-kinds.h"
-#include <memory>
 #include <optional>
 #include <ostream>
-#include <vector>
 
 namespace Fortran::evaluate {
-
-// Intrinsics are distinguished by their names and the characteristics
-// of their arguments, at least for now.
-using IntrinsicProcedure = const char *;  // not an owning pointer
 
 struct CallCharacteristics {
   parser::CharBlock name;
   bool isSubroutineCall{false};
-};
-
-struct SpecificIntrinsic {
-  explicit SpecificIntrinsic(IntrinsicProcedure n) : name{n} {}
-  SpecificIntrinsic(IntrinsicProcedure n, std::optional<DynamicType> &&dt,
-      int r, semantics::Attrs a)
-    : name{n}, type{std::move(dt)}, rank{r}, attrs{a} {}
-  std::ostream &Dump(std::ostream &) const;
-
-  IntrinsicProcedure name;
-  bool isRestrictedSpecific{false};  // if true, can only call it
-  std::optional<DynamicType> type;  // absent if and only if subroutine call
-  int rank{0};
-  semantics::Attrs attrs;  // ELEMENTAL, POINTER
 };
 
 struct SpecificCall {
